04_EndToEndDev/samples/04-05-FunctionApp-MSI/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
PostgreSQL-connector-python
azure.identity